backend/requirements.txt
fastapi
uvicorn
python-dotenv

# AI + LangChain stack
langchain
langchain-core
langchain-openai
langgraph
openai

# Optional: LLM providers
groq

# Data + Utilities
pydantic
requests

# Database (if needed)
pymongo
psycopg2-binary

# File handling / PDF
reportlab
pillow

# ASGI
anyio
starlette
sniffio
